morning-launcher/requirements.txt
requests>=2.28.0
schedule>=1.2.0
python-dotenv>=1.0.0

